For for brokenness with fork() and epoll()
It turns that that if you fork() and have an epoll instance, it
does not create a new epoll copy of the old one.  It just uses
the old one.  So anything you change in the new process will
change the old one, and vice versa.

Add a function to be called on fork that will clean up after
a fork to avoid this issue.

Signed-off-by: Corey Minyard <cminyard@mvista.com>