selector: Fix multi-threading issues
There was an issue with handling events in a multi-threaded program.
If an fd was deleted then a new fd was put in it's place, a select/epoll
return for that fd may have been for the old fd, but could be
processing the new event data.  To fix this, add a delete counter
and if it changes while in select/epoll, don't do any processing.

Also rework the saved events handling so that it works on exceptions
being enabled as well as read.  This simplified the interface a bit
and is cleaner.

Also rename the update call to be "fd", not "epoll".

Signed-off-by: Corey Minyard <cminyard@mvista.com>