Add an option '-8' to the SDR compiler to use 8-bit ASCII encoding only
For some reason the current version of the ipmitool utility doesn't
support ASCII encoding other than 8-bits per character.
This could lead to a gibberish in the sensor name field, if SDR compiler
chose to use another kind of text encoding.
To avoid this add an option '-8' to the SDR compiler to use 8-bit ASCII text
encoding only.

Signed-off-by: Corey Minyard <cminyard@mvista.com>