lanserv: Ack all packet with a sequence number
Not just packets with data, that way you get acks for breaks, CTS
setting, etc.

The spec is unclear on this, it says in one place that only packets with
data will be acked, but it says in another that it must ack each packet
it receives unless it's an ack-only packet.  This will work with both
interpretations, though.

Signed-off-by: Corey Minyard <cminyard@mvista.com>