sol: Rewrite the library code
I started doing serious tests with gensio and all kinds of issues came
up.  The code is convoluted, has too many locks and timers, and is
really messy.

This is a massive simplification.  There is only one main data
structure, one lock, one timer, and no dynamic allocation on data
transfer.  Almost a complete rewrite of most of the code.  The APIs are
unchanged for the most part, except for some semantics.

The biggest change is that the write and set routines that send data can
return EAGAIN, meaning that the queues are full or an operation is still
pending.  The operation should be held and done again when the operation
completes or the data transfer completes.

Signed-off-by: Corey Minyard <cminyard@mvista.com>