sol: Rework how break, ring, etc. operations handle no acks
Some BMCs (probably most) do not acknowledge packets without data, even
if the sequence number field is set.  The spec is duplicitous on the
matter: it says both that only data packets are acked and that all
packets with sequence numbers are acked.

The OpenIPMI SOL code will detect if a BMC acks non-data packets with a
recent change.  This change makes the operations that have this issue
return an error if you set a callback on them and they won't get an ack.
It still does the operation, it just returns a specific error code.

Signed-off-by: Corey Minyard <cminyard@mvista.com>