entity: Handle NULL entity in internal_fru_fetch_done()
In the case the entity is destroyed while doing a FRU fetch, it is
possible to get internal_fru_fetch_done() called with a NULL entity
because the entity lookup will fail.  The entity data still exists
because we get/put it, so pass in the entity in the callback data.

Signed-off-by: Corey Minyard <cminyard@mvista.com>