Fix the name of global event support for sensors
In the spec, there is an event support option for "global disable",
meaning that the sensor cannot be individually enabled and disabled,
only the entire MC can have events enabled and disabled.

Unfortunately, I named the define for this
IPMI_EVENT_SUPPORT_GLOBAL_ENABLE.  It's functionally the same, I suppose,
but add a IPMI_EVENT_SUPPORT_GLOBAL_DISABLE define with the same value,
switch over to it, and leave the old one in for backwards compatibility
purposes.

Signed-off-by: Corey Minyard <cminyard@mvista.com>