lan: Don't set reponse LUN to 2 in bridged messages
That causes the response to be queued in the receive queue, according
to the IPMI spec.  For some reason it works on some systems, but not
others.

Set it to 0 to get the resposne back to the proper place.

Note that ipmitool sets it to 0 in this case, so this should be a safe
change.

Signed-off-by: Corey Minyard <corey@minyard.net>