lanserv/bmc_sensor: Adjust Generator ID
Right now channel number is being unconditionally included in the
Generator ID field. However, as per the SEL Event Records it should
remain 0h depending on what medium/interface the event msg is
received over.

So have this:

 SEL has no entries
 # ipmitool event 1
 Sending SAMPLE event: Temperature - Upper Critical - Going High
 # ipmitool sel get 0x1 | grep "Generator ID"
  Generator ID          : 0041

Instead of:

 # ipmitool event 1
 Sending SAMPLE event: Temperature - Upper Critical - Going High
 # ipmitool sel get 0x2 | grep "Generator ID"
  Generator ID          : f041

As we are at it adjust the msg length of the platform event - as per the
Table 29-4, Platform Event (Event Message) Command, the 8 bytes is
a msg length dedicated for the System Interface which must include
the Generator ID. Case in point, when the event is sent over the
LAN channel, ipmi_sim rejects it due to invalid length of the
request (e.g. ipmitool does not include the extra software ID in the
event data hence sending only 7 bytes):

 $ ipmitool -Ilanplus -Hlocalhost -p4242 -UXXXXX -PXXXXX event 1
 Sending SAMPLE event: Temperature - Upper Critical - Going High
 Platform Event Message command failed: Request data length invalid

Signed-off-by: Michal Berger <michallinuxstuff@gmail.com>
Message-ID: <CACosJgzCym0fGOh4_itgnV5Dk+vOqk=Xy+dtvbmCGcpKCFY5XA@mail.gmail.com>
[I heavily modified this patch from Michal.  Even with his changes,
 the data location and handling of source address/generator ID and
 source LUN were wrong.  I figured I would go ahead and fix all the
 problems.]
Signed-off-by: Corey Minyard <corey@minyard.net>